pyproject.toml
[tool.poetry]
name = "fms"
version = "0.1.0"
description = "symple financial management system"
authors = ["Vitaliy Kozhushko <itvkip@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "^3.8.1"
django = "^5.1.2"
sqlparse = "^0.5.1"
typing-extensions = "^4.12.2"
tzdata = "^2024.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
environs = "^11.0.0"
psycopg2 = "^2.9.9"
pyjwt = "^2.9.0"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^23.0.0"
drf-yasg = "^1.21.7"
django-rest-swagger = "^2.2.0"
celery = "^5.4.0"
redis = "^5.1.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
flake8 = "^7.1.1"
pylint = "^3.3.1"
django-stubs = "^5.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
